Implement simplified g4sbs_tree structure for hcal
Since most detectors share many data structure in common,
here is an example of how to simplify the hcal data
structure.  Each data structure (Cal, ECal, ECalPart) is
responsible for setting up the branches. This requires
a significant smaller number of variables in g4sbs_tree that
will need to be maintained.